fix(cost): include cache tokens and resolve dup-session ingest crash
Real cost was understated by ~75% on long Opus sessions: cache_read_input_tokens
and cache_creation_input_tokens were ignored, so a session with 87M cache reads
showed $50.86 instead of the actual $203.71. Date-versioned model names like
claude-opus-4-5-20251101 also failed exact-match lookup and fell back to Sonnet
rates.

- jsonl_parser._price: prefix match after exact match for date-suffixed names
- jsonl_parser: count cache_write at 125% input rate, cache_read at 10%
- session_summary.tokens: BigInteger so 64-bit totals don't overflow
- ingest: look up existing rows by summary.session_id (not path.stem) so
  continuation/agent files sharing a parent sessionId hit the existing row
  instead of UNIQUE-violating
- Pages.jsx: format token counts as M when >=1M, k otherwise

+++ backend/app/models/session_summary.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from datetime import datetime
 
-from sqlalchemy import DateTime, Float, Integer, String
+from sqlalchemy import BigInteger, DateTime, Float, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.db import Base
@@ -18,7 +18,7 @@ class SessionSummaryRow(Base):
     agent: Mapped[str | None] = mapped_column(String, nullable=True)
     task: Mapped[str | None] = mapped_column(String, nullable=True)
     status: Mapped[str] = mapped_column(String, default="completed")
-    tokens: Mapped[int] = mapped_column(Integer, default=0)
+    tokens: Mapped[int] = mapped_column(BigInteger, default=0)
     cost: Mapped[float] = mapped_column(Float, default=0.0)
     edits: Mapped[int] = mapped_column(Integer, default=0)
     file_path: Mapped[str] = mapped_column(String)
